.gitlab-ci.yml: run and enforce cppcheck
Run cppcheck in CI, currently on the common folder.

The results are kept as an artifact to allow developers to check the
results in detail.

Now that cppcheck uses a compilation database, we can generate it from
the CI configuration file to have more control over the build
configuration.

Add a new hidden job: .with-compilation-database, because other static
analysis tools can make use of the database (clang-tidy for example).

Add a new "cppcheck" job, to run cppcheck with the generated database.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>