.gitlab-ci.yml: add support for other DUTs
Add support for other devices by using the new '-d' option of
run_test_file.sh.

The currently accepted devices are:
"prplmesh": prplmesh with dummy BWL.
"prplmesh-rax40": prplmesh on the rax40 with DWPAL BWL.

By default, "prplmesh" is used.

To be able to use the current prplMesh version on the rax40, it also
needs to be deployed first, so add a call to tools/deploy.sh before
running the tests.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>