framework: fix cppcheck error (shared_ptr dealloc)
When an explicit call to shared_ptr is made when calling
std::vector.push_back(), cppcheck believes the ptr is deallocated, and
should not be returned.

While it's a bug in cppcheck, the additional call to shared_ptr() is
not necessary, as it will create one more temporary ptr which is
immediately copied into the container.

Remove the call to shared_ptr to avoid the useless temporary.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>