common: fix call to virtual function in destructor (cppcheck)
Every bwl makes a call to a virtual function in its
destructor. cppcheck reports it as a warning because when called from
a destructor/constructor, only the function from the same class will
be called, not the derived ones.

Fix it for now by adding the explicitly qualified ID of the function,
to make it clear that we will only call the class's function and not
any derived one.

Another way to fix it would have been to make all the derived classes
final to make it even more clear that there is no derived class to
call the function from. This is however not possible, as
sta_wlan_hal_dummy inherits from base_wlan_hal_dummy which inherits
from base_wlan_hal, so we can't make the function final in
base_clan_hal_dummy (sta_wlan_hal overrides it).

In the future and depending on the implementation of the detach()
function, we might also want to solve it in a different way: split
detach() by using a private function m_detach() that every derived
class implement and that is called from both the destructor and the
virtual detach() function.

Fixes 5 cppcheck issues of severity warning.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>