controller: bml: fix results type
To receive the channel scan results the caller sends an empty initialized
results list, the allocated storage is then used to store the incoming
results, previously this flow caused a `segmentation fault` error and lead
to a crush scenario.

To fix the issue the results variable in the caller function needed to
be changed from `BML_NEIGHBOR_AP**` to `BML_NEIGHBOR_AP*` since the
initialization moved from within the bwl_internal to the caller
function.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>