tools: scripts: deploy_ipk.sh add --proxy option
deploy_ipk.sh is currently used by the CI to copy the built prplmesh ipk
and install it to the target.
It cannot be used by Intel developers since our build server is not
directly connected to the lab setups, but connected via proxy.
Add proxy support to the script so it can be used to deploy ipks over
proxy.

While at it, fix some shellcheck errors in the script and add the
license snippet which was missing.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>