cmake/Findelpp.cmake: remove
Now we don't do find_package(elpp) any more, Findelpp.cmake is no longer
needed.

It might be relevant if we want to be able to use an external
easylogging++ in the future. However, the current Findelpp.cmake is not
appropriate for that at all. So we can just as well remove it.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>