.gitlab-ci.yml: openwrt: remove the deletion of the docker images
Keeping the built docker images on the Gitlab runners is essential to
benefit from the docker cache and avoid rebuilding everything from
scratch every time.

It's however not realistic to keep any image built forever, so what we
did up to now is to only keep the images built on the master branch.

This approach is however problematic when we need to update the
images, because:
- the images for the branch used for the update are not kept, so the
  CI pipelines take much more time to complete until the branch is
  merged.
- once the branch is merged into master, the updated image is the one
  being kept, but the previous ones get deleted. As a consequence, any
  side branch still using the old images will have their pipeline time
  increase a lot because everything is potentially rebuilt from
  scratch.

A possible solution is to check the last time an image was tagged, and
to only remove the ones which haven't been tagged for the last X days.

Depending on the number of docker images to delete, this operation can
be costly in terms of time, so doing it for every single pipeline
doesn't make much sense.

To avoid repeating the operation for every pipeline, it could be done
in a single job ran on schedule, but that in turn means it's not
guaranteed to be run on every Gitlab runners (since at the moment we
have no way to force a job to be picked up by a specific runner).

For these reasons and because we control all the runners we use for
now, it's easier to handle this task on the runners themselves, using
a cron job or a systemd timer for example.

This commit removes the deletion of the docker images entirely, so
each runner will have to make sure it handles the deletion of the
images itself.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>