framework: ieee1905_transport: redundant null pointer check
When the condition (!addr0 && !addr1) was reached, it could never have
been evaluated to true, because both addr0 and addr1 get default
values if they are null.

Move the logging to the beginning of the function to keep the original
intention to warn the user in case both are null.

Fixes 6 cppcheck issues of severity warning.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>